[Bot] Add check for bot_attack_flag for ^attack to prevent accidental bot movements.
Previously if a bot was told to ^attack but was mid-cast, stunned or otherwise distacted, once they were free to attack they would attack the owner's current target which could result in the bot attacking the wrong target.

This will throw a flag on an NPC that bots have been told to ^attack that will last for 10 seconds.

When a bot is flagged to attack, if the owner's current target does not have an attack flag they will not attack the target.