[Bots][Spells] Corrects IsImmune for bots and message outputs for spells failures/partials
Prior to changes, when a bot was doing immune checks before casting a spell it was using player cast logic as if the spell was already being casted which could result in spammed aggro for the bot on immunechecks.

Cleans up and adds message outputs for bots similar to players.
So and so is immune, partial slows, unaffected by lull, etc.