[Anticheat] Adds checks for skipping locked and triggered doors. Prevents casting, pulling and attacking of mobs on opposite sides of locked/triggered doors.
Rule - Map, CheckForLoSCheat, false, "Runs predefined zone checks to check for LoS cheating through doors and such."
Rule - Map, RangeCheckForLoSCheat, 20.0, "Default 20.0. Range to check if one is within range of a door."
Rule - Range, DoChecksForIllusionGlitching, false, "Controls where or not coords are updating and doors are checked to see if a client is trying to illusion or horse through the door."
Rule - Range, UpdateRangeForIllusionChecks, 50.0, "Default 50.0. This is how far one must travel to initiate an update of the client's position for the illusion glitch check."
Rule - Range, UpdateIntervalForIllusionChecks, 250, "Default 250ms. In ms this is how often the position is updated to revert for illusions."